+++ packages/gamut/src/Form/SelectDropdown/elements/index.ts
@@ -6,7 +6,8 @@ export {
   RemoveAllButton,
 } from './multi-value';
 
-export { DropdownButton, onFocus } from './controls';
+export { DropdownButton } from './controls';
+export { onFocus } from '../core/accessibility';
 
 export {
   SelectDropdownContext,
